+++ CHANGELOG.md
@@ -38,6 +38,7 @@ It also supports the new "added" timestamp for songs in the mpd database.
 - Feat: Simplify Last-Played implementation and save it in MessagePack format
 - Feat: "Add random songs/album" is executed in worker thread
 - Feat: Rework structs for better thread safety
+- Feat: Add toggle button to show mpd search expression instead of breadcrumbs
 - Upd: Manual add random songs or albums should ignore the jukebox constraints
 - Upd: libmympdclient 1.0.25